requirements.txt
Django==4.2.17
djangorestframework==3.15.2
drf_social_oauth2==3.1.0
reportlab==4.2.5
python-dotenv==1.0.1
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
psycopg2==2.9.10